fix bad example of storage authz in scope
Motivation:

An existing example contradicts the stated authorisation policy for
`storage.create`.  It is also badly phrased, which may lead to confusion.

Modification:

Rework English prose description of the expected authorisation.

Add an additional example that is rather similar to the existing example
to illustrate the behaviour of `storage.modify`.

Result:

Clearer storage AuthZ examples.

Closes: #31