add definition of a VO
Motivation:

The document makes many references to VOs but neither defining for what
"VO" stands, nor describing what is a virtual organisation.

This is bad because the document makes a destinction between VOs (as
top-level collection of users) and group (as non-top-level collection of
useres), so the definition matters.

Modification:

Modify first paragraph to use "virtual organisation" instead of VO.

Add definition of a VO.

Result:

It becomes easier to understand whether a group of people constitute a
VO.

Closes: #38