remove non-AuthZ, protocol-specific behaviour constraint
Motivation:

The document current places constraints on how the server reacts to
requests that are unrelated to an authorisation decision.  Such
constraints are outside the scope of this document.

Modification:

Remove offending statement.

Result:

The document is more closely focused on AuthN and AuthZ, potentially
improving its maintainability and applicability.

Closes: #35