Document edits not changing the semantic
Some edits to improve the grammar and punctuation, to use the USA English spelling to ease spell-checking (both UK and USA variants were present and some cited documents and technical terms use the USA spelling), and to clarify some sentences (minimal changes).
Changes were kept to the minimum with the goal of not changing the semantics and not adding or removing content.